+++ packages/blade/src/components/Charts/CommonChartComponents/index.ts
@@ -5,7 +5,9 @@ export type {
   ChartTooltipProps,
   ChartLegendProps,
   ChartReferenceLineProps,
+  ChartLabelContentProps,
   DataColorMapping,
+  SecondaryLabelMap,
 } from './types';
 export {
   ChartXAxis,